address review: restore target-dependent integer widening; drop stale T2/T4 comments
- cares_jsc.rs: restore i32::from(hostent.h_addrtype) — h_addrtype is c_short on
  Windows, c_int on POSIX; the compare needs the widen. Gated with
  allow(useless_conversion) so linux-only clippy doesn't sweep it again.
- snapshot.rs: restore u64::from(c_ulong) for line_col_to_byte_offset args —
  c_ulong is u32 on Windows (LLP64). Same allow gate.
- jsc_hooks.rs: import_kind: import_kind → field-init shorthand.
- unlink_command/pack_command/publish_command/init_command/pm_view_command/
  JSONCObject: drop comments describing the removed T2→T4 .into() lift and the
  no-op identity rebinds they justified.

+++ src/runtime/dns_jsc/cares_jsc.rs
@@ -98,7 +98,8 @@ pub fn hostent_with_ttls_to_js_response(
             // so build a sockaddr_in/in6 on the stack and copy through that.
             let addr_string = {
                 // h_addrtype is c_short on Windows, c_int on POSIX; widen for the compare.
-                let address = if hostent.h_addrtype == c_ares::AF::INET6 {
+                #[allow(clippy::useless_conversion)]
+                let address = if i32::from(hostent.h_addrtype) == c_ares::AF::INET6 {
                     // SAFETY: addr points to ≥16 bytes for AF_INET6.
                     let bytes: [u8; 16] = unsafe { *(addr as *const [u8; 16]) };
                     let mut sa6: super::netc::sockaddr_in6 = bun_core::ffi::zeroed();
